MasterManagementServiceInterfaceを削除
